requirements.txt
flask-api==1.1
flask-restful==0.3.7
flask-restplus==0.12.1
matplotlib==3.0.3
pandas==0.24.2
requests==2.21.0
scikit-learn==0.20.3
